+++ parser/src/main/java/io/ticofab/androidgpxparser/parser/task/GpxFetchedAndParsed.java
@@ -1,0 +1,7 @@
+package io.ticofab.androidgpxparser.parser.task;
+
+import io.ticofab.androidgpxparser.parser.domain.Gpx;
+
+public interface GpxFetchedAndParsed {
+    void onGpxFetchedAndParsed(Gpx gpx);
+}