+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Change Log
 ==========
 
+Version 0.1.3 *(2015-09-12)*
+----------------------------
+
+ * Removed Google Play Services dependency
+ * Isolated taks and interface (breaking change)
+
 Version 0.1.1 *(2015-09-12)*
 ----------------------------